camera/thermal/requirements.txt
numpy
opencv-python
requests
urllib3
python-dotenv